requirements.txt
torch==2.2.2
torchvision==0.17.2
torchaudio==2.2.2
opencv-python==4.10.0.84
pillow==10.3.0
numpy==1.26.4
pandas==2.2.2
matplotlib==3.8.4
scikit-learn==1.4.2
labelme==5.3.1
streamlit==1.38.0
tqdm==4.66.4
albumentations==1.4.3
typing-extensions>=4.11.0

